+++ packages/analytics/analytics-utilities/src/types/explore/requests.ts
@@ -1,25 +1,35 @@
 import { makeFilterable } from './util'
-import type { AbsoluteTimeRangeV4, MetricFilterTypesV2 } from './common'
-import { queryableExploreDimensions, type ExploreFilter } from './advanced'
+import type { AbsoluteTimeRangeV4, ExploreFilterTypesV2, RequestFilterTypeEqualsV2, RequestFilterTypeMetricV2, RequestFilterTypeWildcardV2 } from './common'
+import { type ExploreFilter } from './advanced'
 
+// status_code and upstream_status_code are treated as metric filters
 export const queryableRequestDimensions = [
-  ...queryableExploreDimensions,
+  'api_product',
+  'api_product_version',
+  'application',
   'auth_type',
   'client_ip',
+  'consumer',
   'consumer_group',
+  'control_plane',
+  'control_plane_group',
   'country_code',
+  'data_plane_node',
+  'gateway_service',
   'header_host',
   'header_user_agent',
   'http_method',
+  'iso_code',
   'request_id',
+  'request_uri',
   'response_header_content_type',
   'response_source',
-  'request_uri',
+  'route',
   'service_port',
   'service_protocol',
   'sse',
-  'upstream_status_code',
-  'upstream_status_code_grouped',
+  'status_code_grouped',
+  'time',
   'upstream_uri',
   'websocket',
 ] as const
@@ -30,13 +40,26 @@ export const filterableRequestDimensions = makeFilterable(queryableRequestDimens
 
 export type FilterableRequestDimensions = typeof filterableRequestDimensions[number]
 
+export const queryableRequestWildcardDimensions = [
+  'request_uri',
+  'upstream_uri',
+] as const
+
+export type QueryableRequestWildcardDimensions = typeof queryableRequestWildcardDimensions[number]
+
+export const filterableRequestWildcardDimensions = queryableRequestWildcardDimensions
+
+export type FilterableRequestWildcardDimensions = typeof filterableRequestWildcardDimensions[number]
+
 export const queryableRequestMetrics = [
+  'ai_count',
   'latencies_response_ms',
   'latencies_upstream_ms',
   'latencies_kong_gateway_ms',
   'response_body_size',
   'request_body_size',
-  'ai_count',
+  'status_code',
+  'upstream_status_code',
 ] as const
 
 export type QueryableRequestMetrics = typeof queryableRequestMetrics[number]
@@ -45,18 +68,37 @@ export const filterableRequestMetrics = makeFilterable(queryableRequestMetrics)
 
 export type FilterableRequestMetrics = typeof filterableRequestMetrics[number]
 
-export interface RequestDimensionFilter extends Omit<ExploreFilter, 'dimension'> {
-  dimension?: FilterableRequestDimensions
-  field?: FilterableRequestDimensions
+export interface RequestInFilter {
+  type: ExploreFilterTypesV2
+  field: FilterableRequestDimensions | FilterableRequestWildcardDimensions | FilterableRequestMetrics
+  value: (string | number)[]
 }
-
+export interface RequestEqualsFilter {
+  type: RequestFilterTypeEqualsV2
+  field: FilterableRequestDimensions | FilterableRequestWildcardDimensions
+  value: string
+}
 export interface RequestMetricFilter {
-  type: MetricFilterTypesV2
+  type: RequestFilterTypeMetricV2
   field: FilterableRequestMetrics
-  value: number | null
+  value: number
+}
+export interface RequestEmptyFilter {
+  type: RequestFilterTypeEqualsV2
+  field: FilterableRequestDimensions | FilterableRequestWildcardDimensions | FilterableRequestMetrics
+}
+export interface RequestWildcardFilter {
+  type: RequestFilterTypeWildcardV2
+  field: FilterableRequestWildcardDimensions
+  value: string
 }
 
-export type RequestFilter = RequestDimensionFilter | RequestMetricFilter
+export type RequestFilter = ExploreFilter |
+  RequestInFilter |
+  RequestEqualsFilter |
+  RequestMetricFilter |
+  RequestEmptyFilter |
+  RequestWildcardFilter
 
 export const relativeTimeRangeValuesRequestV2 = [
   '15M',