+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.4...@kong-ui-public/analytics-geo-map@0.2.5) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [0.2.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.3...@kong-ui-public/analytics-geo-map@0.2.4) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [0.2.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.2...@kong-ui-public/analytics-geo-map@0.2.3) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
 ## [0.2.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.1...@kong-ui-public/analytics-geo-map@0.2.2) (2024-09-09)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-geo-map