+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.4.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.4.4...@kong-ui-public/analytics-metric-provider@8.4.5) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [8.4.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.4.3...@kong-ui-public/analytics-metric-provider@8.4.4) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [8.4.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.4.2...@kong-ui-public/analytics-metric-provider@8.4.3) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [8.4.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.4.1...@kong-ui-public/analytics-metric-provider@8.4.2) (2024-09-09)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider