+++ packages/analytics/portal-analytics-bridge/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.29](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/portal-analytics-bridge@0.2.28...@kong-ui-public/portal-analytics-bridge@0.2.29) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/portal-analytics-bridge
+
+
+
+
+
+## [0.2.28](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/portal-analytics-bridge@0.2.27...@kong-ui-public/portal-analytics-bridge@0.2.28) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/portal-analytics-bridge
+
+
+
+
+
+## [0.2.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/portal-analytics-bridge@0.2.26...@kong-ui-public/portal-analytics-bridge@0.2.27) (2024-09-11)
+
+**Note:** Version bump only for package @kong-ui-public/portal-analytics-bridge
+
+
+
+
+
 ## [0.2.26](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/portal-analytics-bridge@0.2.25...@kong-ui-public/portal-analytics-bridge@0.2.26) (2024-09-06)
 
 **Note:** Version bump only for package @kong-ui-public/portal-analytics-bridge